requirements.txt
# Backend
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
pgvector>=0.2
neo4j==6.0.3
redis>=5.0
python-dotenv==1.1.1
pydantic==2.12.3

# RAG
langgraph==0.2.61
langchain==0.3.26
langchain-openai==0.3.8
fastapi>=0.109
uvicorn==0.38.0
jinja2>=3.1
langchain-community==0.3.27

# Data Import & Scripts (Added in feature/graphDB)
pandas==2.3.3
openpyxl==3.1.5
requests==2.32.5